Added file validation to both download and populate methods. Merged the two together into the same file.

+++ nasa_download.rb
@@ -1,14 +1,19 @@
 require 'rubygems'
 require 'rest-client'
 
-alpha ||= 'a'
-nasa_url = "http://er.jsc.nasa.gov/seh/"
-nasa_local_filename = "pages/nasa_"
+def nasa_download
+	alpha ||= 'a'
+	nasa_url = "http://er.jsc.nasa.gov/seh/"
+	nasa_local_filename = "pages/nasa_"
 
-while alpha != 'z'.next
-	File.open("#{nasa_local_filename}#{alpha}.html", "w") do |file|
-	   file.write(RestClient.get("#{nasa_url}#{alpha}.html"))
+	while alpha != 'z'.next
+
+		unless File.exist?("#{nasa_local_filename}#{alpha}.html")
+			File.open("#{nasa_local_filename}#{alpha}.html", "w") do |file|
+			   file.write(RestClient.get("#{nasa_url}#{alpha}.html"))
+			end
+		end
+
+		alpha.next!
 	end
-
-	alpha.next!
 end
